Implement custom devserver
The existing ones I've tested {live,live}-server don't work because they depend on the built files being actual files instead of symlinks (which Bazel uses).

So, we customize a devserver that uses stdin as a signal instead as according to ibazel_notify_changed.

However, this isn't ideal since we're always reloading the entire page so maybe we can explore how to retrieve a list of files changed and we run bs.reload on those files directly.